capitulo_8/recursos/ejercicio_3/api/requirements.txt
absl-py==0.9.0
adal==1.2.2
ale-py==0.6.0.dev0
apturl==0.5.2
asn1crypto==1.1.0
astor==0.8.1
astroid==2.3.1
attrs==19.3.0
backcall==0.1.0
binwalk==2.1.1
bleach==3.1.0
bokeh==1.2.0
Brlapi==0.6.6
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
clickclick==1.2.2
cloudpickle==1.2.1
command-not-found==0.3
connexion==2.4.0
contextlib2==0.5.5
cryptography==2.7
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.6
dask==2.1.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.5.0
distributed==2.1.0
distro-info===0.18ubuntu0.18.04.1
ecdsa==0.15
edgetpu==2.11.1
entrypoints==0.3
esptool==2.8
et-xmlfile==1.0.1
filterpy==1.4.5
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==1.4.1
gast==0.3.3
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-pasta==0.1.8
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
graphene==1.4.1
graphene-sqlalchemy==1.1.1
graphql-core==1.1
graphql-relay==0.4.5
grpcio==1.27.1
h5py==2.10.0
HeapDict==1.0.0
httplib2==0.9.2
idna==2.8
impeve==1.0
importlib-metadata==1.3.0
imutils==0.5.3
inflection==0.3.1
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.11
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka==1.3.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kfp==0.1
kiwisolver==1.1.0
kubernetes==8.0.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.2
llvmlite==0.31.0
locket==0.2.0
logilab-common==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==8.0.2
msgpack==0.6.2
mysql-connector-python==8.0.18
nbconvert==5.4.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.7
numba==0.47.0
numpy==1.18.1
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
openapi-spec-validator==0.2.8
opencv-python==4.1.2.30
openpyxl==3.0.3
opt-einsum==3.1.0
packaging==19.2
paho-mqtt==1.5.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.3.4
partd==1.0.0
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.6.0
promise==2.0.2
prompt-toolkit==2.0.9
protobuf==3.11.3
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.4.2
pymacaroons==0.13.0
pymongo==3.9.0
PyNaCl==1.1.2
pyocr==0.7.2
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyqtgraph==0.10.0
pyRFC3339==1.0
pyrsistent==0.15.5
pyserial==3.4
pytesseract==0.3.2
pytest==5.3.2
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.0.1
qtconsole==4.4.3
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
smart-box==0.4
sngular-box==0.3
sortedcontainers==2.1.0
SQLAlchemy==1.1.11
swagger-ui-bundle==0.0.5
system-service==0.3
systemd-python==234
tb-nightly==1.15.0a20190911
tblib==1.4.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf-estimator-nightly==2.0.0.dev2019102608
tf-nightly==1.15.0.dev20190730
toolz==0.9.0
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
typing==3.6.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.6
usb-creator==0.3.3
vboxapi==1.0
virtualenv==16.7.5
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.0
widgetsnbextension==3.4.2
wrapt==1.11.2
xkit==0.0.0
xlrd==1.2.0
zict==1.0.0
zipp==0.6.0
zope.interface==4.3.2
